+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file, and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Added student-set initial conditions per [#159](https://github.com/microsoft/pybryt/issues/159)
+
 ## 0.5.0 - 2022-02-15
 
 * Added complexity unions for combining complexity classes per [#155](https://github.com/microsoft/pybryt/issues/155)